Send garden output to syslog and log files
Following the pattern in CF and Diego releases, send the logs from
garden to syslog in addition to the original log files.

I decided to log with priority user.info and user.error for stdout and
stderr, as they do in the diego-release jobs.

Note that logger will run as root and the log files will still be own
by root.