Tidy up bosh package and submodule sync scripts
* Previously there were a lot of unnecessary entries added to the
`guardian-deps` script, which resulted in redundant files getting added
to the guardian BOSH package.
* This has all been tidied up such that only the dependencies we
require are included in the BOSH package spec.
* Similar changes have been made to the sync-submodule-config script,
and as a result we now include fewer submodules.
* Additionally the github.com/golang/protobuf may have been updated.
* Most of the original confusion seemed to stem from the fact that we
weren't specifying GOOS=linux when running the sync scripts, thus we
were accidentally including an "incorrect" set of dependencies for a
linux runtime (when running scripts from darwin).

[#145185373]

Signed-off-by: Will Martin <wmartin@pivotal.io>