Rootless guardian tests pass on new Concourse
We upgraded concourse from version 3.0.1 to 3.2.1, and the rootless GQTs
began failing because the new{u,g}idmap executables in our docker image
were somehow losing their setuid permissions at test run time.

We aren't sure why this is happening but restoring the setuid in the
test script fixes this, as expected.

[#146366143]