Ensure gdn has no supplementary GIDs
In the rootless+TCP case, we used to use chpst to run the process as
maximus. This had the unwanted side effect of adding maximus to the
supplementary groups, which cannot be dropped by runc when rootless.
This caused all non-root container processes to have root as a
supplementary group, which was a leak of privilege.

[finishes #152439956]

Signed-off-by: Julia Nedialkova <julianedialkova@hotmail.com>