Reduce the size of the droplet in cfbench to ~100mb
At this moment the diego-release that makes oci buildpack apps
use exclusive disk quota is not released yet.
Because of that we can't use bigger droplets, or the app will fail to
start due to the quota be smaller than the app size itself.
Setting the droplet to ~100mb ensures it have enough disk quota to start
and is still within the acceptance parameters.

[finishes #153114195]

Signed-off-by: Tiago Scolari <tscolari@pivotal.io>