Create the grootfs store directory ourselves
If we let BPM create the directory it would be owned by vcap and
permissioned as 700. This means that a rootless garden/grootfs are not
going to be able to access its subdirectories hence container creation
fails.

When we create the directory in the BPM prestart hook we rely on a sane
umask (0022) which would create the directory and set its permissions
to 755.

Co-authored-by: Danail Branekov <danailster@gmail.com>