Set a hard limit for the GrootFS store size
This changes the `thresholder` to set the `store_size_bytes` config
property to a value that is either the same as the GrootFS GC threshold,
or the size of the whole disk when the disk isn't big enough (15GB).

This way we still preserve the old behaviour for users with small disks
who are happy to continuously collect garbage, while making sure that we
don't fill up the disk in all the other cases.

[#165764224]

Co-authored-by: Kieron Browne <kbrowne@pivotal.io>