Change containerd config to use ephemeral state directory
The containerd task plugin stores its shim pipes and sockets for running
containers in `/run/containerd/`. Its state goes in the containerd state
directory. `/run/` is an ephemeral temp directory and is cleared out on
reboot. Containerd expects its state directory to be ephemeral too but
we have been storing it in `/var/vcap/data/containerd/state` which,
though not bosh persistent, is preserved between reboots.

This leads to an issue if a VM running containerd is rebooted while it
has running containers. The state directory will have details of those
containers, but the pipes and sockets will have disappeared. Due to a
bug, containerd will currently wait 100 seconds for a container's shim
socket to appear before marking the container as deleted. With multiple
containers, this means containerd takes many minutes to start up and
means diego cells become unhealthy.

This small change aligns garden's containerd configuration with the
expectations of the containerd code. We also include a migration to copy
the state directory to the new location if garden is upgraded without
recreating its bosh instance.

Issue: https://github.com/cloudfoundry/garden-runc-release/issues/195
Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>