Use isolated filesystem for thresholder tests
Using the / filesystem in concourse leads to flakes. Its available
blocks can vary as other test containers are created on the worker node.

Therefore, we should isolate the tests by using a filesystem that will
not be used by any other process. We do this with a sparse file and
create an ext4 filesystem on it.

Issue: https://garden.ci.cf-app.com/teams/main/pipelines/main/jobs/thresholder/builds/403
Co-authored-by: Mario Nitchev <marionitchev@gmail.com>