Compile tar with musl-gcc
Glibc is still loading some libraries dynamically even though we compile
tar statically. If runtime glibc version (in container rootfs) is
different from the compilation glibc version there are might be
incompatible errors. We observed the following error when using Ubuntu
jammy and cflinuxfs3 during droplet stream-out:

https://sourceware.org/bugzilla/show_bug.cgi?id=27790

The issue is that we use tar on a host compiled in garden-runc-release
with glibc in ubuntu jammy. Then, we use nstar to stream in/out data
from the container. Nstar is switching namespace to container rootfs and
thus using glibc in linuxfs3.

Instead of using glibc we build tar with musl-gcc.

"Some of musl's major advantages over glibc and uClibc/uClibc-ng are its
size, correctness, static linking support, and clean code."

https://www.musl-libc.org/