Update garden/grootfs startup to be compatible with bpm > 1.1.9
Changes to the underlying runc libraries used in BPM changed the
permissions of device control, requiring garden to explicitly set them
at startup within BPM.

When not run inside a container, the `permit_device_control` function
exits without making any changes. As such this should be safe to run
as part of the startup script.