Build pkg-config and iptables on bionic
Resulting binary for when iptables compiled with musl wouldn't
execute. This is due to the fact the pkg-config has to use glib. In this
case, it's using an internal glibc. Instead if we build the two binaries
using ubuntu:bionic, it will continue to work on both Jammy and Bionic.

Reference cloudfoundry/garden-runc-release#233