ci: update guardian-standalone
Fixes:
```
  [FAILED] Set TEST_ROOTFS Env variable
  Expected
      <bool>: false
  to be true
```

There were some changes to the garden-integration-tests that modified
the required environment variables to run the integration test suite.
See this commit for the beginning of that work:
https://github.com/cloudfoundry/garden-integration-tests/commit/d3f3e898f278d51eee4d45b4b23cee89e235d892

Fixes:
```
Ginkgo detected a version mismatch between the Ginkgo CLI and the version of Ginkgo imported by your packages:
  Ginkgo CLI Version:
    2.10.0
  Mismatched package versions found:
    2.11.0 used by garden-integration-tests
```
By invoking ginkgo through raw go (and thus the package version in the go
mod) instead of invoking the ginkgo binary in the container.