Enable guardian-windows to cleanup process state.
This uses similar logic to guardian on Linux, but we defaulted
the behavior (`cleanup_process_dirs_on_wait`) to be enabled.
We believe it was disabled in Linux for Concourse when originally
implemented, as Diego/CF require it to be on. Since Concourse
doesn't use garden on Windows workers, we defaulted this on to
make it easier on the CF side.

[#185836130](https://www.pivotaltracker.com/story/show/185836130)

Signed-off-by: Amelia Downs <adowns@vmware.com>